add TileViewModel and resolve circular dependency issues, re #3559
